refactor(Balance): AND-1216 Replace repeated price and format code with code that uses CryptoCurrency enum. Deprecate some more methods that do not take a CryptoCurrency or CryptoValue
* refactor(Balance): AND-1216 Introduce Precision to cope with the two Eth string formats

+++ app/src/test/java/piuk/blockchain/android/ui/account/AccountEditPresenterTest.kt
@@ -55,7 +55,6 @@ import piuk.blockchain.androidcore.data.metadata.MetadataManager
 import piuk.blockchain.androidcore.data.payload.PayloadDataManager
 import piuk.blockchain.androidcore.utils.PrefsUtil
 import piuk.blockchain.androidcoreui.ui.customviews.ToastCustom
-import java.math.BigDecimal
 import java.math.BigInteger
 import java.util.ArrayList
 import java.util.Arrays
@@ -205,7 +204,7 @@ class AccountEditPresenterTest {
             .thenReturn("USD")
         whenever(sendDataManager.estimateSize(anyInt(), anyInt())).thenReturn(1337)
 
-        whenever(currencyFormatManager.getFormattedSelectedCoinValue(BigDecimal.TEN))
+        whenever(currencyFormatManager.getFormattedSelectedCoinValue(BigInteger.TEN))
             .thenReturn("")
 
         // Act
